# src/main/java/com/vlastolabs/quartzdash/domain/TriggerKey.java
/*
 * Copyright 2026 VlastoLabs Software
 * and other contributors as indicated by the @author tags.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package  com.vlastolabs.quartzdash.domain;

/**
 * Composite key identifying a Quartz trigger by group and name.
 *
 * @param group the trigger group
 * @param name  the trigger name
 */
public record TriggerKey(String group, String name) {

    public static TriggerKey of(String group, String name) {
        return new TriggerKey(group, name);
    }

    public String qualifiedName() {
        return group + "." + name;
    }
}